+++ src/app_error.rs
@@ -16,17 +16,16 @@ pub enum AppError {
 /// Convert errors into strings to display
 impl fmt::Display for AppError {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        let disp = match self {
-            Self::DockerConnect => "Unable to access docker daemon".to_owned(),
-            Self::DockerInterval => "Docker update interval needs to be greater than 0".to_owned(),
-            Self::InputPoll => "Unable to poll user input".to_owned(),
-            Self::Terminal => "Unable to draw to terminal".to_owned(),
-            Self::DockerCommand(s) => format!("Unable to {} container", s),
+        match self {
+            Self::DockerConnect => write!(f, "Unable to access docker daemon"),
+            Self::DockerInterval => write!(f, "Docker update interval needs to be greater than 0"),
+            Self::InputPoll => write!(f, "Unable to poll user input"),
+            Self::Terminal => write!(f, "Unable to draw to terminal"),
+            Self::DockerCommand(s) => write!(f, "Unable to {} container", s),
             Self::MouseCapture(x) => {
                 let reason = if *x { "en" } else { "dis" };
-                format!("Unable to {}able mouse capture", reason)
+                write!(f, "Unbale to {}able mouse capture", reason)
             }
-        };
-        write!(f, "{}", disp)
+        }
     }
 }