+++ src/docker_data/mod.rs
@@ -1,18 +1,19 @@
 use bollard::{
     container::{ListContainersOptions, LogsOptions, StartContainerOptions, Stats, StatsOptions},
+    service::ContainerSummary,
     Docker,
 };
 use futures_util::StreamExt;
 use parking_lot::Mutex;
 use std::{
     collections::HashMap,
-    fmt,
     sync::{
         atomic::{AtomicBool, Ordering},
         Arc,
     },
 };
 use tokio::{sync::mpsc::Receiver, task::JoinHandle};
+use uuid::Uuid;
 
 use crate::{
     app_data::{AppData, DockerControls},
@@ -25,17 +26,25 @@ pub use message::DockerMessage;
 
 #[derive(Debug, Hash, Clone, PartialEq, Eq)]
 enum SpawnId {
-    Stats(String),
+    Stats((String, Binate)),
     Log(String),
 }
 
-impl fmt::Display for SpawnId {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        let disp = match self {
-            Self::Stats(id) => format!("stats::{id}"),
-            Self::Log(id) => format!("logs::{id}"),
-        };
-        write!(f, "{}", disp)
+/// Cpu & Mem stats take twice as long as the update interval to get a value, so will have two being executed at the same time
+/// SpawnId::Stats takes container_id and binate value to enable both cycles of the same container_id to be inserted into the hashmap
+/// Binate value is toggled when all join handles have been spawned off
+#[derive(Debug, Hash, Clone, PartialEq, Eq, Copy)]
+enum Binate {
+    One,
+    Two,
+}
+
+impl Binate {
+    const fn toggle(self) -> Self {
+        match self {
+            Self::One => Self::Two,
+            Self::Two => Self::One,
+        }
     }
 }
 
@@ -48,6 +57,7 @@ pub struct DockerData {
     receiver: Receiver<DockerMessage>,
     spawns: Arc<Mutex<HashMap<SpawnId, JoinHandle<()>>>>,
     timestamps: bool,
+    binate: Binate,
 }
 
 impl DockerData {
@@ -87,6 +97,7 @@ impl DockerData {
         app_data: Arc<Mutex<AppData>>,
         is_running: bool,
         spawns: Arc<Mutex<HashMap<SpawnId, JoinHandle<()>>>>,
+        spawn_id: SpawnId,
     ) {
         let mut stream = docker
             .stats(
@@ -109,14 +120,15 @@ impl DockerData {
 
             let cpu_stats = Self::calculate_usage(&stats);
 
-            let no_bytes = (0, 0);
+            let no_bytes = || (0, 0);
+
             let (rx, tx) = if let Some(key) = some_key {
                 match stats.networks.unwrap_or_default().get(&key) {
-                    Some(data) => (data.rx_bytes.to_owned(), data.tx_bytes.to_owned()),
-                    None => no_bytes,
+                    Some(data) => (data.rx_bytes, data.tx_bytes),
+                    None => no_bytes(),
                 }
             } else {
-                no_bytes
+                no_bytes()
             };
 
             if is_running {
@@ -133,7 +145,7 @@ impl DockerData {
                     .lock()
                     .update_stats(&id, None, None, mem_limit, rx, tx);
             }
-            spawns.lock().remove(&SpawnId::Stats(id.clone()));
+            spawns.lock().remove(&spawn_id);
         }
     }
 
@@ -143,26 +155,28 @@ impl DockerData {
             let docker = Arc::clone(&self.docker);
             let app_data = Arc::clone(&self.app_data);
             let spawns = Arc::clone(&self.spawns);
-            let is_running = *is_running;
             let id = id.clone();
 
-            let key = SpawnId::Stats(id.clone());
-            let spawn_contains_id = spawns.lock().contains_key(&key);
-            let s = tokio::spawn(Self::update_container_stat(
-                docker,
-                id.clone(),
-                app_data,
-                is_running,
-                spawns,
-            ));
-            if !spawn_contains_id {
-                self.spawns.lock().insert(key, s);
-            }
+            let key = SpawnId::Stats((id.clone(), self.binate));
+
+            let spawn_key = key.clone();
+            self.spawns.lock().entry(key).or_insert_with(|| {
+                tokio::spawn(Self::update_container_stat(
+                    docker,
+                    id.clone(),
+                    app_data,
+                    *is_running,
+                    spawns,
+                    spawn_key,
+                ))
+            });
         }
+        self.binate = self.binate.toggle();
     }
 
     /// Get all current containers, handle into ContainerItem in the app_data struct rather than here
     /// Just make sure that items sent are guaranteed to have an id
+    /// Will ignore any container that contains `oxker` as an entry point
     pub async fn update_all_containers(&mut self) -> Vec<(bool, String)> {
         let containers = self
             .docker
@@ -173,24 +187,33 @@ impl DockerData {
             .await
             .unwrap_or_default();
 
-        let mut output = vec![];
-        // iter over containers, to only send ones which have an id, as use id for identification throughout!
-        containers
+        let mut output = containers
             .iter()
-            .filter(|i| i.id.is_some())
-            .for_each(|c| output.push(c.clone()));
+            .filter_map(|f| match f.id {
+                Some(_) => {
+                    if f.command.as_ref().map_or(false, |c| c.contains("oxker")) {
+                        None
+                    } else {
+                        Some(f.clone())
+                    }
+                }
+                None => None,
+            })
+            .collect::<Vec<ContainerSummary>>();
 
-        self.app_data.lock().update_containers(&output);
+        self.app_data.lock().update_containers(&mut output);
 
         let current_sort = self.app_data.lock().get_sorted();
         self.app_data.lock().set_sorted(current_sort);
 
+        // Just get the containers that are currently running, or being restarted, no point updating info on paused or dead containers
         output
             .iter()
             .filter_map(|i| {
                 i.id.as_ref().map(|id| {
                     (
-                        i.state.as_ref().unwrap_or(&String::new()) == "running",
+                        i.state == Some("running".to_owned())
+                            || i.state == Some("restarting".to_owned()),
                         id.clone(),
                     )
                 })
@@ -241,60 +264,59 @@ impl DockerData {
             let app_data = Arc::clone(&self.app_data);
             let spawns = Arc::clone(&self.spawns);
             let key = SpawnId::Log(id.clone());
-            let s = tokio::spawn(Self::update_log(
-                docker, id, timestamps, 0, app_data, spawns,
-            ));
-
-            self.spawns.lock().insert(key, s);
+            self.spawns.lock().insert(
+                key,
+                tokio::spawn(Self::update_log(
+                    docker, id, timestamps, 0, app_data, spawns,
+                )),
+            );
         }
     }
 
+    /// Update all cpu_mem, and selected container log (if a log update join_handle isn't currently being executed)
     async fn update_everything(&mut self) {
         let all_ids = self.update_all_containers().await;
         let optional_index = self.app_data.lock().get_selected_log_index();
         if let Some(index) = optional_index {
+            // this could be neater
             let id = self.app_data.lock().containers.items[index].id.clone();
-
             let key = SpawnId::Log(id.clone());
-            let running = self.spawns.lock().contains_key(&key);
 
-            if !running {
+            self.spawns.lock().entry(key).or_insert_with(|| {
                 let since = self.app_data.lock().containers.items[index].last_updated as i64;
                 let docker = Arc::clone(&self.docker);
                 let timestamps = self.timestamps;
-
                 let app_data = Arc::clone(&self.app_data);
                 let spawns = Arc::clone(&self.spawns);
-                let s = tokio::spawn(Self::update_log(
+                tokio::spawn(Self::update_log(
                     docker, id, timestamps, since, app_data, spawns,
-                ));
-                self.spawns.lock().insert(key, s);
-            }
+                ))
+            });
         };
-
         self.update_all_container_stats(&all_ids).await;
     }
 
     /// Animate the loading icon
-    async fn loading_spin(&mut self) -> JoinHandle<()> {
+    async fn loading_spin(&mut self, loading_uuid: Uuid) -> JoinHandle<()> {
         let gui_state = Arc::clone(&self.gui_state);
         tokio::spawn(async move {
             loop {
                 tokio::time::sleep(std::time::Duration::from_millis(100)).await;
-                gui_state.lock().next_loading();
+                gui_state.lock().next_loading(loading_uuid);
             }
         })
     }
 
     /// Stop the loading_spin function, and reset gui loading status
-    fn stop_loading_spin(&mut self, handle: &JoinHandle<()>) {
+    fn stop_loading_spin(&mut self, handle: &JoinHandle<()>, loading_uuid: Uuid) {
         handle.abort();
-        self.gui_state.lock().reset_loading();
+        self.gui_state.lock().remove_loading(loading_uuid);
     }
 
     // Initialize docker container data, before any messages are received
     async fn initialise_container_data(&mut self) {
-        let loading_spin = self.loading_spin().await;
+        let loading_uuid = Uuid::new_v4();
+        let loading_spin = self.loading_spin(loading_uuid).await;
 
         let all_ids = self.update_all_containers().await;
         self.update_all_container_stats(&all_ids).await;
@@ -312,7 +334,7 @@ impl DockerData {
             self.initialised = self.app_data.lock().initialised(&all_ids);
         }
         self.app_data.lock().init = true;
-        self.stop_loading_spin(&loading_spin);
+        self.stop_loading_spin(&loading_spin, loading_uuid);
     }
 
     /// Handle incoming messages, container controls & all container information update
@@ -320,57 +342,58 @@ impl DockerData {
         while let Some(message) = self.receiver.recv().await {
             let docker = Arc::clone(&self.docker);
             let app_data = Arc::clone(&self.app_data);
+            let loading_uuid = Uuid::new_v4();
             match message {
                 DockerMessage::Pause(id) => {
-                    let loading_spin = self.loading_spin().await;
-                    docker.pause_container(&id).await.unwrap_or_else(|_| {
+                    let loading_spin = self.loading_spin(loading_uuid).await;
+                    if docker.pause_container(&id).await.is_err() {
                         app_data
                             .lock()
                             .set_error(AppError::DockerCommand(DockerControls::Pause));
-                    });
-                    self.stop_loading_spin(&loading_spin);
+                    };
+                    self.stop_loading_spin(&loading_spin, loading_uuid);
                 }
                 DockerMessage::Restart(id) => {
-                    let loading_spin = self.loading_spin().await;
-                    docker
-                        .restart_container(&id, None)
-                        .await
-                        .unwrap_or_else(|_| {
-                            app_data
-                                .lock()
-                                .set_error(AppError::DockerCommand(DockerControls::Restart));
-                        });
-                    self.stop_loading_spin(&loading_spin);
+                    let loading_spin = self.loading_spin(loading_uuid).await;
+                    if docker.restart_container(&id, None).await.is_err() {
+                        app_data
+                            .lock()
+                            .set_error(AppError::DockerCommand(DockerControls::Restart));
+                    };
+                    self.stop_loading_spin(&loading_spin, loading_uuid);
                 }
                 DockerMessage::Start(id) => {
-                    let loading_spin = self.loading_spin().await;
-                    docker
+                    let loading_spin = self.loading_spin(loading_uuid).await;
+                    if docker
                         .start_container(&id, None::<StartContainerOptions<String>>)
                         .await
-                        .unwrap_or_else(|_| {
-                            app_data
-                                .lock()
-                                .set_error(AppError::DockerCommand(DockerControls::Start));
-                        });
-                    self.stop_loading_spin(&loading_spin);
+                        .is_err()
+                    {
+                        app_data
+                            .lock()
+                            .set_error(AppError::DockerCommand(DockerControls::Start));
+                    };
+                    self.stop_loading_spin(&loading_spin, loading_uuid);
                 }
                 DockerMessage::Stop(id) => {
-                    let loading_spin = self.loading_spin().await;
-                    docker.stop_container(&id, None).await.unwrap_or_else(|_| {
+                    let loading_spin = self.loading_spin(loading_uuid).await;
+                    if docker.stop_container(&id, None).await.is_err() {
                         app_data
                             .lock()
                             .set_error(AppError::DockerCommand(DockerControls::Stop));
-                    });
-                    self.stop_loading_spin(&loading_spin);
+                    };
+                    self.stop_loading_spin(&loading_spin, loading_uuid);
                 }
                 DockerMessage::Unpause(id) => {
-                    let loading_spin = self.loading_spin().await;
-                    docker.unpause_container(&id).await.unwrap_or_else(|_| {
+                    let loading_spin = self.loading_spin(loading_uuid).await;
+                    if docker.unpause_container(&id).await.is_err() {
                         app_data
                             .lock()
                             .set_error(AppError::DockerCommand(DockerControls::Unpause));
-                    });
-                    self.stop_loading_spin(&loading_spin);
+                    };
+                    // loading sping take uuid to remove
+                    // stop_loading_sping(uuid)
+                    self.stop_loading_spin(&loading_spin, loading_uuid);
                     self.update_everything().await;
                 }
                 DockerMessage::Update => self.update_everything().await,
@@ -405,6 +428,7 @@ impl DockerData {
                 spawns: Arc::new(Mutex::new(HashMap::new())),
                 timestamps: args.timestamp,
                 is_running,
+                binate: Binate::One,
             };
             inner.initialise_container_data().await;