+++ src/ui/mod.rs
@@ -57,13 +57,13 @@ pub async fn create_ui(
     )
     .await;
 
-    disable_raw_mode().unwrap_or(());
+    disable_raw_mode()?;
     execute!(
         terminal.backend_mut(),
         LeaveAlternateScreen,
         DisableMouseCapture
     )?;
-    terminal.show_cursor().unwrap_or(());
+    terminal.show_cursor()?;
 
     if let Err(err) = res {
         println!("{}", err);
@@ -85,23 +85,21 @@ async fn run_app<B: Backend + Send>(
 
     // Check for docker connect errors before attempting to draw the gui
     let e = app_data.lock().get_error();
-    if let Some(error) = e {
-        if let AppError::DockerConnect = error {
-            let mut seconds = 5;
-            loop {
-                if seconds < 1 {
-                    is_running.store(false, Ordering::SeqCst);
-                    break;
-                }
-                if terminal
-                    .draw(|f| draw_blocks::error(f, &AppError::DockerConnect, Some(seconds)))
-                    .is_err()
-                {
-                    return Err(AppError::Terminal);
-                }
-                tokio::time::sleep(std::time::Duration::from_secs(1)).await;
-                seconds -= 1;
+    if let Some(AppError::DockerConnect) = e {
+        let mut seconds = 5;
+        loop {
+            if seconds < 1 {
+                is_running.store(false, Ordering::SeqCst);
+                break;
             }
+            if terminal
+                .draw(|f| draw_blocks::error(f, &AppError::DockerConnect, Some(seconds)))
+                .is_err()
+            {
+                return Err(AppError::Terminal);
+            }
+            tokio::time::sleep(std::time::Duration::from_secs(1)).await;
+            seconds -= 1;
         }
     } else {
         let mut now = Instant::now();
@@ -109,7 +107,7 @@ async fn run_app<B: Backend + Send>(
             if terminal.draw(|f| ui(f, &app_data, &gui_state)).is_err() {
                 return Err(AppError::Terminal);
             }
-            if crossterm::event::poll(input_poll_rate).unwrap_or_default() {
+            if crossterm::event::poll(input_poll_rate).unwrap_or(false) {
                 if let Ok(event) = event::read() {
                     if let Event::Key(key) = event {
                         sender